requirements.txt
asyncio
pandas-gbq
httpx
fastapi
uvicorn